LoongArch: KVM: Repair the restart pressure test virtual machine hang dead problem
Upstream: no

After the virtual restart, exintc does not correctly clear
the estat register and qemu does not clear the estat register.
Estat data sent from qemu is merged instead of overwritten on kvm.

In order to solve this problem,
the following changes were made to this patch:

1.exintc correctly clears zeros when the VM restarts
2. estat on the qemu side clears ESTAT correctly when the VM restarts
3. When the estat register data on the qemu side is sent to the kvm,
the data needs to be overwritten instead of merged

+++ arch/loongarch/kvm/intc/ipi.c
@@ -540,7 +540,7 @@ int kvm_loongarch_register_ipi_device(void)
 int kvm_loongarch_reset_ipi(struct kvm_vcpu *vcpu)
 {
 	struct ipi_state *s = &vcpu->arch.ipi_state;
-	u8 offset, size;
+	unsigned long offset, size;
 	u8 *pstart;
 
 	if (!s)
